pyproject.toml
[project]
name = "repl-nix-discordrankreporter"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.4.0",
]
